+++ tests/products.test.js
@@ -1962,7 +1962,7 @@ describe('Products', () => {
               _id
               status
               ... on ConfigurableProduct {
-                simulatedPriceRange(vectors: [{ key: "COLOR", value: "red" }]) {
+                simulatedPriceRange(vectors: [{ key: "color-variant", value: "color-variant-red" }]) {
                   _id
                   minPrice {
                     _id
@@ -1991,13 +1991,13 @@ describe('Products', () => {
         minPrice: {
           isTaxable: true,
           isNetPrice: false,
-          amount: 500000,
+          amount: 1500000,
           currency: 'CHF',
         },
         maxPrice: {
           isTaxable: true,
           isNetPrice: false,
-          amount: 1500000,
+          amount: 10000000,
           currency: 'CHF',
         },
       });
@@ -2118,7 +2118,7 @@ describe('Products', () => {
               _id
               status
               ... on ConfigurableProduct {
-                catalogPriceRange(vectors: [{ key: "COLOR", value: "red" }]) {
+                catalogPriceRange(vectors: [{ key: "color-variant", value: "color-variant-red" }]) {
                   _id
                   minPrice {
                     _id
@@ -2147,13 +2147,13 @@ describe('Products', () => {
         minPrice: {
           isTaxable: true,
           isNetPrice: false,
-          amount: 500000,
+          amount: 1500000,
           currency: "CHF",
         },
         maxPrice: {
           isTaxable: true,
           isNetPrice: false,
-          amount: 1500000,
+          amount: 10000000,
           currency: "CHF",
         },
       });