+++ tests/seeds/products.js
@@ -565,6 +565,12 @@ export const ProxyProduct = {
         },
         productId: ProxyPlanProduct1._id,
       },
+      {
+        vector: {
+          'color-variant': 'color-variant-blue',
+        },
+        productId: ProxyPlanProduct2._id,
+      },
     ],
   },
 };